Fix id argument in role delete method
This was changed to $roleId in 87452c49b1c29e2a1310cb5ea4ad3b35d80c72b6, because the internal code actually used a variable called $roleId which was simply constructed by casting $id to a string before type hinting as added in 22b95460cd22d7b0dadf9b1f2d224550a978dc44. However this change results in a fatal error because the RoleControllerProvider defines the parameter name as {id} and not {roleId}. And changing it to {roleId} in the ControllerProvider results in CORS issues, because other HTTP methods on the same path (get, patch) use {id} so this confuses Silex it seems. We could also fix it by making all of these controller methods use $roleId instead of $id but that's a bigger change.