Make the console app throw errors instead of catching them
By default the Symfony ConsoleApplication will catch exceptions, print an error in the output, and then exit.
This is now changed so we can handle the exception ourselves by logging it in Sentry and then printing the same output.
There's also no need to re-throw it in the cli app, since we already show it by printing the output through the ConsoleOutput which adds better formatting than PHP would do for unhandled exceptions.