Only send unusual exceptions to Sentry
Instead of starting by sending every exception to Sentry as soon as one is caught by Silex, we should only send those that don't have a specific handler or are at least not caught by the user themselves (DataValidationException, EntityNotFoundException, etc). This prevents a flood of unfixable errors in Sentry.