Add easy way to get boolean from query parameters
Query parameters are normally always parsed as strings. Because casting a string to a boolean in PHP sometimes has unexpected results, it's useful to put this logic in a re-usable class that works with the PSR-7 Request's query parameter array. We can also easily add support for a default value this way.