Refactor Offer::importLabels() to take visibility changes into account
When a label is imported that is already on the event/place with a different visibility, the label with the old visibility will be removed (first) and then the same label with the new visibility will be added. This way all event listeners downstream can handle the visibility change without introducing a new "LabelVisibilityChanged" or "LabelUpdated" for events/places.