Fix PHPStan error
The type was correct, but PHPStan complained that the query() method on that type does not expect an argument. Which is technically true because the method definition does not define any arguments. But the method uses func_get_args() to get one or more args from the method call, so this method call is in fact valid. The easiest way to make PHPStan ignore is to make it impossible for it to know the type of $db