Differentiate between unknown file size and empty file size
getSize() may return 0, in which case the file is empty. In that case we return an ApiProblem. But in case getSize() returns null, the file size could probably not be determined due to a server issue. So in that case we return a RuntimeException again, so we notice it in Sentry