Improve performance of canonical sorting methods (see details commit)
We currently first loop trough all places, and then sort the arrays based on some field.
By directly comparing event counts, we avoid the need to sort the array, reducing complexity, and also lowering the RAM usage because we don't build a tmp array to sort and then throw away..

Because sorting in PHP uses quicksort, which is O(n log n). This means this change reduces the complexity to a linear O(n)